feat: streamline init prompts
